\name{cfma-package}
\alias{cfma-package}
\alias{cfma}
\docType{package}
\title{
Causal Functional Mediation Analysis
}
\description{
cfma package performs causal functional mediation analysis (CFMA) for functional treatment, functional mediator, and functional outcome. This package includes two functional mediation model type: (1) a concurrent mediation model and (2) a historical influence mediation model.
}
\details{
\tabular{ll}{
	Package: \tab cfma \cr
	Type: \tab Package \cr
	Version: \tab 1.0 \cr
	Date: \tab 2018-05-16 \cr
	License: \tab GPL (>=2) \cr
}
}
\author{
Yi Zhao <zhaoyi1026@gmail.com> and Xi Luo <xi.rossi.luo@gmail.com> and Martin Lindquist <mal2053@gmail.com> and Brian Caffo <bcaffo@gmail.com>

Maintainer: Yi Zhao <zhaoyi1026@gmail.com>
}
\keyword{ package }

